refactor: 삭제하지 않는 어노테이션 삭제
