+++ src/main/java/wooteco/subway/domain/strategy/fare/distance/DistanceFareStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.strategy.fare.basic;
+package wooteco.subway.domain.strategy.fare.distance;
 
 public interface DistanceFareStrategy {
