+++ src/main/java/wooteco/subway/domain/strategy/fare/distance/SecondDistanceFareStrategy.java
@@ -1,24 +1,19 @@
-package wooteco.subway.domain.strategy.fare.basic;
+package wooteco.subway.domain.strategy.fare.distance;
 
-import org.springframework.stereotype.Component;
-
-@Component
-public class AdvancedDistanceFareStrategy implements DistanceFareStrategy {
+public class SecondDistanceFareStrategy implements DistanceFareStrategy {
 
     private static final int BASIC_FARE = 1250;
     private static final int FIRST_SURCHARGE = 100;
-    private static final int SECOND_SURCHARGE = 100;
 
     private static final int BASIC_DISTANCE = 10;
     private static final int INTERVAL_ONE = 50;
     private static final int NO_FARE_DISTANCE = 0;
 
     private static final double INTERVAL_ONE_SURCHARGE_UNIT = 5.0;
-    private static final double INTERVAL_TWO_SURCHARGE_UNIT = 8.0;
 
     @Override
     public int calculateFare(int distance) {
-        return BASIC_FARE + calculateIntervalOne(distance) + calculateIntervalTwo(distance);
+        return BASIC_FARE + calculateIntervalOne(distance);
     }
 
     private int calculateIntervalOne(int distance) {
@@ -26,10 +21,5 @@ public class AdvancedDistanceFareStrategy implements DistanceFareStrategy {
                 Math.max(distance - BASIC_DISTANCE, NO_FARE_DISTANCE), INTERVAL_ONE - BASIC_DISTANCE)
                 / INTERVAL_ONE_SURCHARGE_UNIT)
                 * FIRST_SURCHARGE;
-    }
-
-    private int calculateIntervalTwo(int distance) {
-        return (int) Math.ceil(Math.max(distance - INTERVAL_ONE, NO_FARE_DISTANCE) / INTERVAL_TWO_SURCHARGE_UNIT)
-                * SECOND_SURCHARGE;
     }
 }
